2024-2/cnn_python_streamlit/requirements.txt
streamlit
streamlit-drawable-canvas
torch
torchvision
pillow